add plugin IDs
