+++ webviz_subsurface/plugins/_parameter_response_correlation/__init__.py
@@ -1,0 +1,1 @@
+from ._plugin import ParameterResponseCorrelation
