+++ webviz_subsurface/plugins/_parameter_response_correlation/_plugin_ids.py
@@ -1,0 +1,25 @@
+class PluginIds:
+    # pylint: disable=too-few-public-methods
+    class Stores:
+        # pylint: disable=too-few-public-methods
+        # Controls
+        SELECTED_ENSEMBLE = "selected-ensemble"
+        SELECTED_RESPONSE = "selected-response"
+        SELECTED_CORRELATION_METHOD = "selected-correlation-method"
+        SELECTED_RESPONSE_AGGREGATION = "slected-response-aggregation"
+        SELECTED_CORRELATION_CUTOFF = "selected-correlation-cutoff"
+        SELECTED_MAX_NUMBER_PARAMETERS = "selected-max-number-parameters"
+        # Filter
+        SELECTED_ZONE = "selected-zone"
+        SELECTED_REGION = "selected-region"
+        SELECTED_PARAMETERS = "selected-parameters"
+
+    class SharedSettings:
+        # pylint: disable=too-few-public-methods
+        FILTER = "filter"
+
+    class ResponseID:
+        # pylint: disable=too-few-public-methods
+        GROUP_NAME = "Response Analyzer"
+
+        RESPONSE_CHART = "response-chart"
